Taking into account feedbacks
