+++ centreon/config/symfony_serializer/TimePeriod/Day.yaml
@@ -1,8 +1,6 @@
 Core\TimePeriod\Domain\Model\Day:
   attributes:
     day:
-      type: string
       groups: ['TimePeriod:Read']
     timeRange:
-      type: string
       groups: ['TimePeriod:Read']
